+++ backend/src/baserow/contrib/database/plugins.py
@@ -18,7 +18,7 @@ from .fields.field_types import (
 class DatabasePlugin(Plugin):
     type = "database"
 
-    def user_created(self, user, group, group_invitation):
+    def user_created(self, user, group, group_invitation, template):
         """
         This method is called when a new user is created. We are going to create a
         database, table, view, fields and some rows here as an example for the user.
@@ -27,7 +27,7 @@ class DatabasePlugin(Plugin):
         # If the user created an account in combination with a group invitation we
         # don't want to create the initial data in the group because data should
         # already exist.
-        if group_invitation:
+        if group_invitation or template:
             return
 
         core_handler = CoreHandler()